rename file to index
